+++ NEWS.md
@@ -5,6 +5,7 @@
 * Allow to pass a list of options to the `contextMenu` argument
 * The value of `jstreeOutput` is now updated when the user 
 creates/deletes/renames/pastes a node
+* Added new Shiny value `outputId_move' triggered when a node is moved
 
 
 # jsTreeR 0.1.0 (2020-08-26)